// src/shared/types/serverTypes.ts
export type Pagination = {
  pageSize?: number;
  pageNumber?: number;
};

export type Sorting = {
  type: 'ASC' | 'DESC';
  field: 'id' | 'createdAt' | 'updatedAt' | 'name';
};

export type PaginationWithTotal = Pagination & {
  total: number;
};

export type GetPageResult<T> = {
  data: T[];
  pagination: PaginationWithTotal;
  sorting: Sorting;
};

export type DateInterval = {
  gte?: string; // от - дата в виде строки new Date().toISOString() 2023-09-19T10:37:16.389+00:00
  lte?: string; // до - дата в виде строки new Date().toISOString() 2023-09-19T10:37:16.389+00:00
};

export type CommonFilters = {
  pagination?: Pagination;
  createdAt?: DateInterval;
  updatedAt?: DateInterval;
  sorting?: Sorting;
};